Task2-3/docker/data_generating/requirements.txt
Faker==30.3.0
psycopg2-binary==2.9.9
